+++ lib/modules/model/ModelsRegister.ts
@@ -11,6 +11,8 @@ import { MeasureDefinition } from "../measure";
 import {
   AssetModelContent,
   DeviceModelContent,
+  GroupModelContent,
+  LocaleDetails,
   MeasureModelContent,
   MetadataDetails,
   MetadataGroups,
@@ -30,6 +32,7 @@ export class ModelsRegister {
   private context: PluginContext;
   private assetModels: AssetModelContent[] = [];
   private deviceModels: DeviceModelContent[] = [];
+  private groupModels: GroupModelContent[] = [];
   private measureModels: MeasureModelContent[] = [];
 
   private get sdk() {
@@ -45,6 +48,7 @@ export class ModelsRegister {
     await Promise.all([
       this.load("asset", this.assetModels),
       this.load("device", this.deviceModels),
+      this.load("group", this.groupModels),
       this.load("measure", this.measureModels),
     ]);
 
@@ -76,6 +80,7 @@ export class ModelsRegister {
     metadataDetails: MetadataDetails = {},
     metadataGroups: MetadataGroups = {},
     tooltipModels: TooltipModels = {},
+    locales: { [valueName: string]: LocaleDetails } = {},
   ) {
     if (Inflector.pascalCase(model) !== model) {
       throw new PluginImplementationError(
@@ -96,6 +101,7 @@ export class ModelsRegister {
     this.assetModels.push({
       asset: {
         defaultMetadata,
+        locales,
         measures,
         metadataDetails,
         metadataGroups,
@@ -156,8 +162,48 @@ export class ModelsRegister {
     });
   }
 
+  /**
+   * Registers a group model.
+   *
+   *
+   * @param engineGroup - The engine group name.
+   * @param model - The name of the group model, which must be in PascalCase.
+   * @param metadataMappings - The metadata mappings for the model, defaults to an empty object.
+   * @param defaultMetadata - The default metadata values for the model, defaults to an empty object.
+   * @param metadataDetails - Optional detailed metadata descriptions, localizations and definition.
+   * @param metadataGroups - Optional groups for organizing metadata, with localizations.
+   * @throws PluginImplementationError if the model name is not in PascalCase.
+   */
+  registerGroup(
+    engineGroup: string,
+    model: string,
+    metadataMappings: MetadataMappings = {},
+    defaultMetadata: JSONObject = {},
+    metadataDetails: MetadataDetails = {},
+    metadataGroups: MetadataGroups = {},
+  ) {
+    if (Inflector.pascalCase(model) !== model) {
+      throw new PluginImplementationError(
+        `Group model "${model}" must be PascalCase`,
+      );
+    }
+
+    // Construct and push the new group model to the groupModels array
+    this.groupModels.push({
+      engineGroup,
+      group: {
+        defaultMetadata,
+        metadataDetails,
+        metadataGroups,
+        metadataMappings,
+        model,
+      },
+      type: "group",
+    });
+  }
+
   registerMeasure(type: string, measureDefinition: MeasureDefinition) {
-    const { validationSchema, valuesMappings, valuesDetails } =
+    const { locales, validationSchema, valuesMappings, valuesDetails } =
       measureDefinition;
     if (validationSchema) {
       try {
@@ -172,6 +218,7 @@ export class ModelsRegister {
 
     this.measureModels.push({
       measure: {
+        locales,
         type,
         validationSchema,
         valuesDetails,