+++ tests/scenario/modules/devices/action-attach-engine.test.ts
@@ -31,6 +31,8 @@ describe("DevicesController:attachEngine", () => {
     ).resolves.toMatchObject({
       _source: {
         engineId: "engine-kuzzle",
+        lastMeasuredAt: null,
+        measures: {},
         _kuzzle_info: {
           author: "-1",
         },